fix warning
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/branches/syslog_aiwi@2088 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/lluf/lluf_utils.ss/src/SystemLog.cpp
@@ -272,7 +272,7 @@ private:
 
         std::string logStr = log.str();
 
-        const int LOG_MAX_SIZE = 1024;
+        const size_t LOG_MAX_SIZE = 1024;
 
         // RFC 3164 says that we must not send messages larger that 1024 bytes.
         if (logStr.size() > LOG_MAX_SIZE)
